- Refeactor model_non_res_default
